requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# RL and ML Dependencies
torch>=2.0.0
stable-baselines3[extra]>=2.0.0
numpy
pandas
gymnasium
tensorboard
matplotlib
seaborn
imageio
pyyaml